pyproject.toml
[project]
name = "ddpm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers[training]>=0.31.0",
    "einops>=0.8.0",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pyyaml>=6.0.2",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.0",
    "wandb>=0.18.7",
]
